Show bottom cubes on volumes
When a volume is active in the combat area inspector, the blocks are faded by
distance. This helps visualize large structures but for small ones, like a
two-block tower, it causes the farthest block to disappear and makes it look
like there's a bug.

This commit changes the fading so that the farthest block never quite
disappears entirely and turns off fading entirely for small 2-block structures.

Signed-off-by: EchKode <48565771+echkode@users.noreply.github.com>