Allow mod to be created in alternate directory
New mods were always being created in the default source folder. This commits
adds a dropdown so you can add a custom project folder to the manager settings
and then pick that location for the mod when it is created.

This commit also adds an option to break the link between the mod ID and the
name of its source directory. That lets you namespace your mod ID (which you
should do if you make your mod public) and also use short directory names.