Add cross-platform compare for YAML files
On export, a bulk file compare operation is performed between the config tree
in the SDK and the copy of that tree in the mod. The comparison routine was
comparing files byte-by-byte which breaks the export on Linux if the SDK
configs have Windows EOL bytes.

This commit implements a dirty workaround that pulls aside YAML files and
compares them line by line after stripping any EOL bytes.