Add button to import localization edits
I got a pull request on a public mod to add another language. After merging the
PR and pulling it down, I had the new localization edits in my project
directory. I added a new button to the utility foldout to slurp in these edits
and add them to the project.